Menu working on all pages
