+++ RepairShop-Android/app/src/main/AndroidManifest.xml
@@ -14,15 +14,21 @@
 
         <activity
             android:name=".MainActivity"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.RepairShopAndroid.NoActionBar">
+            android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
 
-        <activity android:name=".BookAppointment" android:label="@string/bookAppointment" />
+        <activity
+            android:name=".BookAppointment"
+            android:label="@string/bookAppointment">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
 
     </application>
