+++ RepairShop-Android/app/src/main/java/ca/mcgill/ecse321/repairshop/BookAppointment.java
@@ -2,14 +2,12 @@ package ca.mcgill.ecse321.repairshop;
 
 import android.os.Bundle;
 
-import androidx.appcompat.app.AppCompatActivity;
-
-public class BookAppointment extends AppCompatActivity {
+public class BookAppointment extends BaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.fragment_second);
+        setContentView(R.layout.activity_book);
     }
 
 }
